Updating imports and removing two halo_quantities that were never fully implemented.
